+++ src/views/auth/register_view.vue
